+++ src/components/TooltipWrapper/utilities.ts
@@ -79,10 +79,11 @@ export function getAlteredVerticalBarPosition(
 
   if (newPosition.horizontal === TooltipHorizontalOffset.Left) {
     const left = getLeftPosition(x, props);
-    x = left.value;
 
     if (left.wasOutsideBounds) {
       newPosition.horizontal = TooltipHorizontalOffset.Right;
+    } else {
+      x = left.value;
     }
   }