+++ src/hooks/tests/useDataForHorizontalChart.test.tsx
@@ -127,4 +127,56 @@ describe('useDataForHorizontalChart()', () => {
       expect(data.longestLabel).toStrictEqual({negative: 0, positive: 0});
     });
   });
+
+  it('allNumbers will not filter out values of 0', () => {
+    function TestComponent() {
+      const data = useDataForHorizontalChart({
+        ...MOCK_PROPS,
+        data: [
+          {
+            name: 'Group 1',
+            data: [
+              {value: 0, key: 'Label 01'},
+              {value: 0, key: 'Label 02'},
+              {value: 0, key: 'Label 03'},
+            ],
+          },
+        ],
+      });
+
+      return <span data-data={`${JSON.stringify(data)}`} />;
+    }
+
+    const result = mount(<TestComponent />);
+
+    const data = JSON.parse(result.domNode?.dataset.data ?? '');
+
+    expect(data.allNumbers).toStrictEqual([0, 0, 0]);
+  });
+
+  it('areAllNegative returns false when all numbers are 0', () => {
+    function TestComponent() {
+      const data = useDataForHorizontalChart({
+        ...MOCK_PROPS,
+        data: [
+          {
+            name: 'Group 1',
+            data: [
+              {value: 0, key: 'Label 01'},
+              {value: 0, key: 'Label 02'},
+              {value: 0, key: 'Label 03'},
+            ],
+          },
+        ],
+      });
+
+      return <span data-data={`${JSON.stringify(data)}`} />;
+    }
+
+    const result = mount(<TestComponent />);
+
+    const data = JSON.parse(result.domNode?.dataset.data ?? '');
+
+    expect(data.areAllNegative).toStrictEqual(false);
+  });
 });